requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
aiohttp==3.8.3
    # via
    #   hikari
    #   lavaplayer
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.2
    # via aiohttp
attrs==22.2.0
    # via
    #   aiohttp
    #   hikari
charset-normalizer==2.1.1
    # via aiohttp
ciso8601==2.3.0
    # via hikari
colorlog==6.7.0
    # via hikari
dataclass-wizard==0.22.2
    # via -r requirements.in
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
hikari[speedups]==2.0.0.dev113
    # via
    #   -r requirements.in
    #   hikari-lightbulb
hikari-lightbulb==2.3.1
    # via -r requirements.in
idna==3.4
    # via yarl
lavaplayer==1.0.10a0
    # via -r requirements.in
multidict==6.0.4
    # via
    #   aiohttp
    #   hikari
    #   yarl
psutil==5.9.4
    # via -r requirements.in
uvloop==0.17.0
    # via -r requirements.in
yarl==1.8.2
    # via aiohttp
